requirements.txt
aiohttp
aiopg
python-yate
pyyaml
sqlalchemy<2.0.0
redis[hiredis]